created control in init.ts
