+++ src/views/main.ts
@@ -1,9 +1,3 @@
 import { GameStart } from "./index.js";
 
-function greet() {
-  var el = GameStart("fefef");
-  return el;
-}
-
-var worm = greet();
-console.log(worm);
+GameStart();
